Update test.js file for a For loop
Any While loops or For loops are used to repeat certain parts of code for us, so that we don't have to keep repeating ourselves in the code itself.  The loop will check a certain condition for us, and, while that condition is true, the loop repeats the code for us.  The Index Variable goes up, each time around the loop.  The Incrementer adds one to the Index Variable each time the loop goes around.  If the incrementer is missing, then the loop cannot stop.  The index variable turns into an infinite loop that crashes our computer.  So please remember to include the incrementer in both while loops and for loops.