Update test.js file to use for loop
This is a practical example of a For loop that uses the loop to create an array of all the anchor tags in the HTML document.  The loop cycles through the elements of the anchor tags by adding a dynamic class name of "links" that depends on their position in the array, then on the page.